+++ src/plugins/grpc/grpcReflectionMetaDataHandler.ts
@@ -2,8 +2,7 @@ import { log } from '../../io';
 import * as models from '../../models';
 import * as utils from '../../utils';
 import * as grpc from '@grpc/grpc-js';
-import { Client } from 'grpc-reflection-js';
-import { fromJSON } from '@grpc/proto-loader';
+import { GrpcReflection } from 'grpc-js-reflection-client';
 import { loadPackageDefinition } from '@grpc/grpc-js';
 import { GrpcRequest, isGrpcRequest } from './grpcRequest';
 
@@ -40,7 +39,7 @@ export async function loadGrpcProtoDefinitionsWithReflection(
   host: string,
   request: GrpcRequest
 ): Promise<Record<string, models.ProtoDefinition>> {
-  const reflectionClient = new Client(host, request.channelCredentials || grpc.credentials.createInsecure());
+  const reflectionClient = new GrpcReflection(host, request.channelCredentials || grpc.credentials.createInsecure());
 
   const serviceNames = await reflectionClient.listServices();
   log.debug('grpc reflection', serviceNames);
@@ -48,8 +47,10 @@ export async function loadGrpcProtoDefinitionsWithReflection(
   const protoDefinitions: Record<string, models.ProtoDefinition> = {};
   for (const serviceName of serviceNames) {
     if (!serviceName.startsWith('grpc.reflection')) {
-      const symbol = await reflectionClient.fileContainingSymbol(serviceName);
-      const packageDefinition = fromJSON(symbol);
+      const descriptor = await reflectionClient.getDescriptorBySymbol(serviceName);
+
+      const packageDefinition = descriptor.getPackageDefinition();
+
       const grpcObject = loadPackageDefinition(packageDefinition);
 
       protoDefinitions[serviceName] = {